+++ tests/unit/test_handle_mgr.py
@@ -7,7 +7,7 @@ from bigxml.handle_mgr import HandleMgr
 def test_no_handle():
     hmgr = HandleMgr()
     with pytest.raises(RuntimeError):
-        hmgr.handle(0)
+        hmgr.iter_from(0)
 
 
 def test_handle():
@@ -17,9 +17,9 @@ def test_handle():
     hmgr.set_handle(handle)
     handle.assert_not_called()
 
-    hmgr.handle(0)
+    hmgr.iter_from(0)
     handle.assert_called_once_with(0)
 
     with pytest.raises(RuntimeError):
-        hmgr.handle(1)
+        hmgr.iter_from(1)
     handle.assert_called_once()